+++ rust-evaluator/src/evaluation/scalars/bindings.rs
@@ -188,7 +188,7 @@ impl ScalarEvaluationEnvironment for ResolvingEnvironment<'_, '_> {
         &self,
         element_id: &str,
         property: &str,
-        target_source_order: usize,
+        target_source_order: f64,
         property_type: &ScalarType,
     ) -> ScalarEvaluation {
         lookup_geometry_property(
@@ -196,7 +196,7 @@ impl ScalarEvaluationEnvironment for ResolvingEnvironment<'_, '_> {
             element_id,
             property,
             target_source_order,
-            Some(self.source_order),
+            Some(self.source_order as f64),
             property_type,
         )
     }
@@ -215,7 +215,7 @@ impl ScalarEvaluationEnvironment for ResolvingEnvironment<'_, '_> {
             point_key,
             property,
             target_source_order,
-            Some(self.source_order),
+            Some(self.source_order as f64),
             property_type,
         )
     }
@@ -227,7 +227,7 @@ impl ScalarEvaluationEnvironment for ResolvingEnvironment<'_, '_> {
         super::geometry_builtin_runtime::GeometryBuiltinRuntimeTarget,
         super::geometry_builtin_runtime::GeometryBuiltinRuntimeError,
     > {
-        resolve_geometry_builtin_target(self.state, self.source_order, target)
+        resolve_geometry_builtin_target(self.state, self.source_order as f64, target)
     }
 }
